requirements.txt
alembic==1.0.7
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
